Fixing so that tests run
