+++ src/profile/profilePane.dom.ts
@@ -1,14 +1,5 @@
-import { Node, IndexedFormula, NamedNode } from 'rdflib'
-import { Namespaces } from 'solid-namespace'
 import { DataBrowserContext } from 'pane-registry'
-
-export function getLabel (subject: Node, kb: IndexedFormula, ns: Namespaces) {
-  var types = kb.findTypeURIs(subject)
-  if (types[ns.foaf('Person').uri] || types[ns.vcard('Individual').uri]) {
-    return 'Your Profile'
-  }
-  return 'Edit your profile' // At the moment, just allow on any object. Like home pane
-}
+import { NamedNode } from 'rdflib'
 
 export function paneDiv (
   context: DataBrowserContext,
