+++ src/trustedApplications/trustedApplications.test.ts
@@ -1,10 +1,15 @@
 /* eslint-env jest */
-const $rdf = require('rdflib')
-const ns = require('solid-namespace')($rdf)
-const {
-  getStatementsToDelete,
-  getStatementsToAdd
-} = require('./trustedApplicationsUtils')
+import * as $rdf from 'rdflib'
+import solidNamespace from 'solid-namespace'
+import { generateRandomString, getStatementsToDelete, getStatementsToAdd } from './trustedApplications.utils'
+
+const ns = solidNamespace($rdf)
+
+describe('generateRandomString', () => {
+  it('generates a random string five characters long', () => {
+    expect(generateRandomString().length).toBe(5)
+  })
+})
 
 describe('getStatementsToDelete', () => {
   it('should return an empty array when there are no statements', () => {
@@ -17,7 +22,7 @@ describe('getStatementsToDelete', () => {
   })
 
   it('should return all statements for the given origin', () => {
-    const mockStore = $rdf.graph()
+    const mockStore = $rdf.graph() as any // @@ TODO Remove casting
     const mockApplication = $rdf.sym('https://app.example')
     const mockOrigin = $rdf.sym('https://origin.example')
     const mockProfile = $rdf.sym('https://profile.example#me')
@@ -35,7 +40,7 @@ describe('getStatementsToDelete', () => {
   })
 
   it('should not return statements for a different origin', () => {
-    const mockStore = $rdf.graph()
+    const mockStore = $rdf.graph() as any // @@ TODO Remove casting
     const mockApplication = $rdf.sym('https://app.example')
     const mockOrigin = $rdf.sym('https://origin.example')
     const mockProfile = $rdf.sym('https://profile.example#me')
@@ -44,7 +49,7 @@ describe('getStatementsToDelete', () => {
     mockStore.add(mockProfile, ns.acl('trustedApp'), mockApplication)
 
     const statementsToDelete = getStatementsToDelete(
-      $rdf.lit('A different origin'),
+      ($rdf.lit as any)('A different origin'), // @@ TODO Remove casting
       mockProfile,
       mockStore,
       ns
@@ -58,7 +63,7 @@ describe('getStatementsToAdd', () => {
   it('should return all required statements to add the given permissions for a given origin', () => {
     const mockOrigin = $rdf.sym('https://origin.example')
     const mockProfile = $rdf.sym('https://profile.example#me')
-    const modes = [ns.acl('Read'), ns.acl('Write')]
+    const modes = [ns.acl('Read').value, ns.acl('Write').value]
 
     const statementsToAdd = getStatementsToAdd(
       mockOrigin,
